==> src/00_get_data.R <==
## The purpose of this script is to
## 1. Discuss the data availability

# 1. Data availability ---------------------------------------------------------

# Dissolved oxygen and temperature data were collected from November 2019  
# through May 2022 continuously with high-frequency data except during periods
# of calibration.

# Nutrient data (TP, SRP, NOx, NH4, TN, DOC, C:N, sediment C:N, ash-free 
# dry mass) were collected monthly from May to October of 2021. 
# For full methods on analyses, see: Culpepper et al.,____ 
# doi: 

# O-18 isotope data were collected as in Boggart et al., 2017
# doi: 10.1007/s10452-021-09939-1
